Among these 1st file is VolumeDetection.py
VolumeDetection.py would result in desired output but would combine all the similar food items. If there are 2 apples and 1 orange in an image, then the output shall be combined calories present in 2 apples. Whereas in VolumeDetection1.py, individual food items' calories are displayed. food_cnn.py is for training and food_predict.py is for testing. Values.py is helpful in printing the most probable food item.